+++ libraries/session/src/main/java/androidx/media3/session/MediaControllerImplBase.java
@@ -388,6 +388,11 @@ import org.checkerframework.checker.nullness.qual.NonNull;
   @Override
   public void play() {
     if (!isPlayerCommandAvailable(Player.COMMAND_PLAY_PAUSE)) {
+      Log.w(
+          TAG,
+          "Calling play() omitted due to COMMAND_PLAY_PAUSE not being available. If this play"
+              + " command has started the service for instance for playback resumption, this may"
+              + " prevent the service from being started into the foreground.");
       return;
     }
 
@@ -522,6 +527,13 @@ import org.checkerframework.checker.nullness.qual.NonNull;
   @Override
   public void setPlayWhenReady(boolean playWhenReady) {
     if (!isPlayerCommandAvailable(Player.COMMAND_PLAY_PAUSE)) {
+      if (playWhenReady) {
+        Log.w(
+            TAG,
+            "Calling play() omitted due to COMMAND_PLAY_PAUSE not being available. If this play"
+                + " command has started the service for instance for playback resumption, this may"
+                + " prevent the service from being started into the foreground.");
+      }
       return;
     }